correction4 tout les type en double

+++ src/model/Titanic.java
@@ -10,11 +10,11 @@ public class Titanic extends IPoint{
 	
 	    
 	    @CsvBindByName(column = "PassengerId")
-	   public int PassengerId;
+	   public double PassengerId;
 	    @CsvBindByName(column = "Survived")
-	   public int Survived;
+	   public double Survived;
 	    @CsvBindByName(column = "Pclass")
-	   public int Pclass;
+	   public double Pclass;
 	   @CsvBindByName(column = "Name")
 	   public String Name;
 	   @CsvBindByName(column = "Sex")
@@ -22,9 +22,9 @@ public class Titanic extends IPoint{
 	   @CsvBindByName(column = "Age")
 	   public double Age;
 	   @CsvBindByName(column = "SibSp")
-	   public int SibSp;
+	   public double SibSp;
 	   @CsvBindByName(column = "Parch")
-	   public int Parch;
+	   public double Parch;
 	   @CsvBindByName(column = "Ticket")
 	   public String Ticket;
 	   @CsvBindByName(column = "Fare")
@@ -52,7 +52,7 @@ public class Titanic extends IPoint{
 	        this.Embarked = Embarked; 
 	    }*/
 
-		public int getPassengerId() {
+		public double getPassengerId() {
 			return PassengerId;
 		}
 
@@ -60,19 +60,19 @@ public class Titanic extends IPoint{
 			PassengerId = passengerId;
 		}
 
-		public int getSurvived() {
+		public double getSurvived() {
 			return Survived;
 		}
 
-		public void setSurvived(int survived) {
+		public void setSurvived(double survived) {
 			Survived = survived;
 		}
 
-		public int getPclass() {
+		public double getPclass() {
 			return Pclass;
 		}
 
-		public void setPclass(int pclass) {
+		public void setPclass(double pclass) {
 			Pclass = pclass;
 		}
 
@@ -96,23 +96,23 @@ public class Titanic extends IPoint{
 			return Age;
 		}
 
-		public void setAge(int age) {
+		public void setAge(double age) {
 			Age = age;
 		}
 
-		public int getSibSp() {
+		public double getSibSp() {
 			return SibSp;
 		}
 
-		public void setSibSp(int sibSp) {
+		public void setSibSp(double sibSp) {
 			SibSp = sibSp;
 		}
 
-		public int getParch() {
+		public double getParch() {
 			return Parch;
 		}
 
-		public void setParch(int parch) {
+		public void setParch(double parch) {
 			Parch = parch;
 		}
